mostly working

+++ internal/exchanges/exchange.go
@@ -1,11 +1,11 @@
 package exchanges
 
 type Exchange interface {
-	Buy(instrument string, price float32, units float32) (OrderResponse, error)
-	Sell(instrument string, price float32, units float32) (OrderResponse, error)
+	Buy(instrument string, price float32, units float64) (OrderResponse, error)
+	Sell(instrument string, price float32, units float64) (OrderResponse, error)
 }
 
 type OrderResponse interface {
 	Price() float32
-	Units() float32
+	Units() float64
 }
